feat(core): Implement product management system
Adds product controller, repository, and view.
Enables product creation, retrieval, and filtering.

+++ src/main/java/com/devtalles/proyect/products/model/ProductRepository.java
@@ -90,7 +90,7 @@ public class ProductRepository {
 
     }
 
-    public void filterByPrice(Long price) throws ProductException {
+    public void filterByPrice(double price) throws ProductException {
         if(products.isEmpty()){
             throw new ProductException("Product list is empty.");
         }
